% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EarthPoints.R
\docType{package}
\name{EarthPoints}
\alias{EarthPoints}
\alias{EarthPoints-package}
\title{EarthPoints: Compute shots on the Earth}
\description{
Convert to different coordinate system and compute different shots
}
