% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL2UTM2MGRS.R
\name{LL2UTM2mgrs}
\alias{LL2UTM2mgrs}
\title{LL2UTM2MGRS}
\usage{
LL2UTM2mgrs(lat, lng)
}
\arguments{
\item{lat}{numeric latitude}

\item{lng}{numeric longitude}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
 \item Latitude
 \item Longitude
 \item Easting
 \item Northing
 \item Zone
 \item Hemisphere
 \item MGRS
}
}
\description{
Convert Lat Long into MGRS through UTM
}
\examples{
LL2UTM2MGRS(-30.4,-40.8)

}
\author{
Ken Harmon <harmkenn@gmail.com>
}
