% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGRS2UTM2LL.R
\name{MGRS2UTM2LL}
\alias{MGRS2UTM2LL}
\title{MGRS2UTM2LL}
\usage{
MGRS2UTM2LL(MGRS)
}
\arguments{
\item{MGRS}{String of 15 characters of Military Grid Reference}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
 \item MGRS
 \item Easting
 \item Northing
 \item Zone
 \item Hemisphere
 \item Latitude
 \item Longitude
}
}
\description{
Convert MGRS into Lat Long through UTM
}
\examples{
MGRS2UTM2LL("11SNV3000010000")

}
\author{
Ken Harmon <harmkenn@gmail.com>
}
