% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/northdes.R
\docType{data}
\name{northdes}
\alias{northdes}
\title{Northing codes for Earth}
\format{A data frame of 405 rows and 5 columns
\describe{
 \item{hem}{Character Value "N" or "S"}
 \item{gze}{Gridzone number "Even" or "Odd"}
 \item{gzl}{Gridzone letter}
 \item{gsnl}{Gridsquare Northing Letter}
 \item{gsnn}{Gridsquare Northing Number}
}}
\source{
I built this based off MGRS on our Earth
}
\usage{
northdes
}
\description{
A dataset containing all north code combinations for MGRS data
}
\keyword{datasets}
