% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_LL.R
\name{polar_ll}
\alias{polar_ll}
\title{polar_LL}
\usage{
polar_ll(lat1d, lng1d, dist, shotd)
}
\arguments{
\item{lat1d}{numeric launch latitude in degrees}

\item{lng1d}{numeric launch longitude in degrees}

\item{dist}{numeric distance in km}

\item{shotd}{numeric bearing in degrees}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
 \item Launch Latitude
 \item Launch Longitude
 \item Distance in KM
 \item Launch Bearing
 \item Landing Latitude
 \item Landing Longitude
 \item Landing Bearing
 \item Midpoint Latitude
 \item Midpoint Longitude
}
}
\description{
Compute a polar shot from a point and a vector
}
\examples{
polar_LL(-30,-40,10,30)

}
\author{
Ken Harmon <harmkenn@gmail.com>
}
