% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shot_LL.R
\name{shot_ll}
\alias{shot_ll}
\title{shot_ll}
\usage{
shot_ll(lat1d, lng1d, lat2d, lng2d)
}
\arguments{
\item{lat1d}{numeric Launch latitude}

\item{lng1d}{numeric Launch longitude}

\item{lat2d}{numeric Landing latitude}

\item{lng2d}{numeric Landing longitude}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
 \item Launch Latitude
 \item Launch Longitude
 \item Landing Latitude
 \item Landing Longitude
 \item Distance in KM
 \item Launch Bearing
 \item Landing Bearing
 \item Midpoint Latitude
 \item Midpoint Longitude
}
}
\description{
Compute the shot from one point to another
}
\examples{
shot_ll(-30,-40,40,30)

}
\author{
Ken Harmon <harmkenn@gmail.com>
}
