% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTM2MGRS2LL.R
\name{UTM2MGRS2LL}
\alias{UTM2MGRS2LL}
\title{UTM2MGRS2LL}
\usage{
UTM2MGRS2LL(Easting, Northing, gzn, hem)
}
\arguments{
\item{Easting}{numeric 6 digit Easting}

\item{Northing}{numeric 7 digit Northing}

\item{gzn}{numeric 2 digit zone}

\item{hem}{character Hemisphere "N" or "S"}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
 \item Easting
 \item Northing
 \item Zone
 \item Hemisphere
 \item Latitude
 \item Longitude
 \item MGRS
}
}
\description{
Convert UTN into MGRS and Lat Long
}
\examples{
UTM2MGRS2LL(530000,3910000,11,"N")

}
\author{
Ken Harmon <harmkenn@gmail.com>
}
